requirements.txt
google-adk
google-cloud-aiplatform>=1.60.0